Auth

+++ controllers/api/posts.js
@@ -14,7 +14,11 @@ function createPost(req, res, next) {
 function deletePost(req, res, next) {
     Post.findById(req.params.id)
         .then(post => {
-            return post.deleteOne()
+            if (post.owner.equals(req.user._id)) {
+                return post.deleteOne()
+            } else {
+                res.sendStatus(401)
+            }   
         })
         .then(() => res.sendStatus(204))
         .catch(next)
@@ -34,8 +38,11 @@ function indexPost (req, res, next){
 function updatePost (req, res, next) {
     Post.findById(req.params.id)
         .then(post => {
-            
-            return post.updateOne(req.body)
+            if (post.owner.equals(req.user._id)) {
+                return post.updateOne(req.body)
+            } else {
+                res.sendStatus(401)
+            }
         })
         .then(() => res.sendStatus(204))
         .catch(next)
